% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krband.R
\name{krband}
\alias{krband}
\title{Calculate the band of Katsevich-Ramdas on TDC's list of discoveries}
\usage{
krband(thresholds, labels, gamma, c = 0.5, lambda = 0.5, interpolate = TRUE)
}
\arguments{
\item{thresholds}{A vector of rejection thresholds resulting from TDC.}

\item{labels}{A vector of (ordered) labels resulting from TDC.}

\item{gamma}{The confidence level of the band. Typical values include
\code{gamma = 0.05} or \code{gamma = 0.01}.}

\item{c}{Determines the ranks of the target score that are considered
winning. Defaults to \code{c = 0.5} for TDC.}

\item{lambda}{Determines the ranks of the target score that are
considered losing. Defaults to \code{lambda = 0.5} for TDC.}

\item{interpolate}{A boolean indicating whether the bands should be
interpolated. Offers a slight boost in performance at the cost of computing
power. Defaults to \code{TRUE}.}
}
\value{
A vector of upper prediction bounds on the FDP for each rejection
threshold.
}
\description{
In TDC, each hypothesis is attached to a winning score and a label (1 for a
target win and -1 for a decoy win). \code{krband()} assumes that the
hypotheses are ordered in decreasing order of winning scores (breaking
ties at random). In particular, the argument \code{labels} is ordered
according to this rule.
}
\examples{
set.seed(123)
thresholds <- c(250, 500, 750, 1000)
labels <- c(
  rep(1, 250),
  sample(c(1, -1), size = 250, replace = TRUE, prob = c(0.9, 0.1)),
  sample(c(1, -1), size = 250, replace = TRUE, prob = c(0.5, 0.5)),
  sample(c(1, -1), size = 250, replace = TRUE, prob = c(0.1, 0.9))
)
gamma <- 0.05
krband(thresholds, labels, gamma)
}
